requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cssselect==1.1.0
cssselect2==0.4.1
defusedxml==0.7.1
fake-useragent==0.1.11
idna==2.10
importlib-metadata==2.1.1
lxml==4.6.3
parse==1.19.0
Pillow==8.1.2
pycparser==2.20
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
requests==2.25.1
requests-html==0.10.0
six==1.15.0
soupsieve==2.2.1
tinycss2==1.1.0
tqdm==4.59.0
urllib3==1.26.4
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
zipp==3.4.1
